refactor(bridge): extract status protocol
